backend/requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bleach==4.1.0
branca==0.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.1.13
django-allauth==0.45.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-swagger==2.0.7
django-silk==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
entrypoints==0.3
folium==0.12.1
gprof2dot==2021.2.21
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.3
notebook==6.4.4
numpy==1.17.3
oauthlib==3.1.1
olefile==0.46
openapi-codec==1.3.2
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.5.0
